requirements.txt
Django==2.1.7
djangorestframework==3.9.1
django-rest-auth==0.9.3
django-rest-auth[with_social]==0.9.3
Pillow==5.4.1
psycopg2==2.7.7
PyNaCl==1.2.1
